add "host" field in json output
The "node" field is very inconvenient since a user specifies list of
nodes and then if they wants to parse json output and associate results
with original list of addresses they needs to parse "node" field because
it looks like `[User.Name@ip-address:port]` but a user specified just ip
addresses.

Therefore, we add new field "host" in JSON output which is the same as
specified by user in stdin or in --host flag.